+++ packages/wm/src/containers/commands/mod.rs
@@ -9,7 +9,6 @@ mod resize_tiling_container;
 mod set_focused_descendant;
 mod toggle_tiling_direction;
 mod wrap_in_split_container;
-mod get_containers_at_pos;
 
 pub use attach_container::*;
 pub use detach_container::*;
@@ -22,4 +21,3 @@ pub use resize_tiling_container::*;
 pub use set_focused_descendant::*;
 pub use toggle_tiling_direction::*;
 pub use wrap_in_split_container::*;
-pub use get_containers_at_pos::*;
